Add block_by_hash macro
The code for Ethereum/Bitcoin `block_by_hash` is identical except the
bitcoin block cache field.  Change this field to `block_cache` to mirror
the Ethereum field and add a macro to implement `block_by_hash`.
