+++ cnd/src/btsieve/bitcoin/cache.rs
@@ -15,13 +15,16 @@ use tokio::sync::Mutex;
 pub struct Cache<C> {
     pub connector: C,
     #[derivative(Debug = "ignore")]
-    pub cache: Arc<Mutex<LruCache<Hash, Block>>>,
+    pub block_cache: Arc<Mutex<LruCache<Hash, Block>>>,
 }
 
 impl<C> Cache<C> {
     pub fn new(connector: C, capacity: usize) -> Cache<C> {
-        let cache = Arc::new(Mutex::new(LruCache::new(capacity)));
-        Cache { connector, cache }
+        let block_cache = Arc::new(Mutex::new(LruCache::new(capacity)));
+        Cache {
+            connector,
+            block_cache,
+        }
     }
 }
 
@@ -35,7 +38,7 @@ where
     fn latest_block(
         &mut self,
     ) -> Box<dyn Future<Item = Self::Block, Error = anyhow::Error> + Send + 'static> {
-        let cache = Arc::clone(&self.cache);
+        let cache = Arc::clone(&self.block_cache);
         let mut connector = self.connector.clone();
 
         let future = async move {
@@ -56,43 +59,4 @@ where
     }
 }
 
-impl<C> BlockByHash for Cache<C>
-where
-    C: BlockByHash<Block = Block, BlockHash = Hash> + Clone,
-{
-    type Block = Block;
-    type BlockHash = Hash;
-
-    fn block_by_hash(
-        &self,
-        block_hash: Self::BlockHash,
-    ) -> Box<dyn Future<Item = Self::Block, Error = anyhow::Error> + Send + 'static> {
-        let connector = self.connector.clone();
-        let cache = Arc::clone(&self.cache);
-        Box::new(Box::pin(block_by_hash(connector, cache, block_hash)).compat())
-    }
-}
-
-async fn block_by_hash<C>(
-    connector: C,
-    cache: Arc<Mutex<LruCache<Hash, Block>>>,
-    block_hash: Hash,
-) -> anyhow::Result<Block>
-where
-    C: BlockByHash<Block = Block, BlockHash = Hash> + Clone,
-{
-    if let Some(block) = cache.lock().await.get(&block_hash) {
-        log::trace!("Found block in cache: {:x}", block_hash);
-        return Ok(block.clone());
-    }
-
-    let block = connector.block_by_hash(block_hash.clone()).compat().await?;
-    log::trace!("Fetched block from connector: {:x}", block_hash);
-
-    // We dropped the lock so at this stage the block may have been inserted by
-    // another thread, no worries, inserting the same block twice does not hurt.
-    let mut guard = cache.lock().await;
-    guard.put(block_hash, block.clone());
-
-    Ok(block)
-}
+impl_block_by_hash!();
