+++ cnd/src/btsieve/mod.rs
@@ -1,6 +1,8 @@
 #![warn(rust_2018_idioms)]
 #![forbid(unsafe_code)]
 
+#[macro_use]
+pub mod block_by_hash;
 pub mod bitcoin;
 pub mod ethereum;
